pyproject.toml
[project]
name = "pyleucht"
version = "0.1.0"
authors = [
    { name = "Paul Wichern" }
]

dependencies = [
    "pygame~=2.6.0",
    "spidev~=3.8",
    "RPi.GPIO~=0.7.1",
]
